*: add a new directory BaseFile
