+++ dbms/src/BaseFile/PosixWritableFile.cpp
@@ -12,10 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+#include <BaseFile/PosixWritableFile.h>
 #include <Common/Exception.h>
 #include <Common/ProfileEvents.h>
 #include <Common/TiFlashMetrics.h>
-#include <Encryption/PosixWritableFile.h>
 #include <fcntl.h>
 #include <unistd.h>
