+++ dbms/src/BaseFile/PosixWriteReadableFile.h
@@ -14,48 +14,34 @@
 
 #pragma once
 
+#include <BaseFile/RateLimiter.h>
+#include <BaseFile/WriteReadableFile.h>
 #include <Common/CurrentMetrics.h>
-#include <Encryption/RateLimiter.h>
-#include <Encryption/WritableFile.h>
-
-#include <string>
+#include <common/types.h>
 
 namespace CurrentMetrics
 {
-extern const Metric OpenFileForWrite;
-}
-
-#ifndef O_DIRECT
-#define O_DIRECT 00040000
-#endif
+extern const Metric OpenFileForReadWrite;
+} // namespace CurrentMetrics
 
 namespace DB
 {
-class PosixWritableFile : public WritableFile
+class PosixWriteReadableFile : public WriteReadableFile
 {
 public:
-    PosixWritableFile(
-        const std::string & file_name_,
+    PosixWriteReadableFile(
+        const String & file_name_,
         bool truncate_when_exists_,
         int flags,
         mode_t mode,
-        const WriteLimiterPtr & write_limiter_ = nullptr);
+        const WriteLimiterPtr & write_limiter_ = nullptr,
+        const ReadLimiterPtr & read_limiter_ = nullptr);
 
-    ~PosixWritableFile() override;
-
-    ssize_t write(char * buf, size_t size) override;
+    ~PosixWriteReadableFile() override;
 
     ssize_t pwrite(char * buf, size_t size, off_t offset) const override;
 
-    off_t seek(off_t offset, int whence) const override;
-
-    std::string getFileName() const override { return file_name; }
-
     int getFd() const override { return fd; }
-
-    void open() override;
-
-    void close() override;
 
     bool isClosed() const override { return fd == -1; }
 
@@ -63,17 +49,18 @@ public:
 
     int ftruncate(off_t length) override;
 
-    void hardLink(const std::string & existing_file) override;
+    ssize_t pread(char * buf, size_t size, off_t offset) const override;
+
+    String getFileName() const override { return file_name; }
+
+    void close() override;
 
 private:
-    void doOpenFile(bool truncate_when_exists_, int flags, mode_t mode);
+    CurrentMetrics::Increment metric_increment{CurrentMetrics::OpenFileForReadWrite, 0};
 
-private:
-    // Only add metrics when file is actually added in `doOpenFile`.
-    CurrentMetrics::Increment metric_increment{CurrentMetrics::OpenFileForWrite, 0};
-    std::string file_name;
-    int fd = -1;
+    String file_name;
+    int fd;
     WriteLimiterPtr write_limiter;
+    ReadLimiterPtr read_limiter;
 };
-
 } // namespace DB
