+++ dbms/src/Encryption/EncryptedWritableFile.h
@@ -14,8 +14,8 @@
 
 #pragma once
 
+#include <BaseFile/WritableFile.h>
 #include <Encryption/AESCTRCipherStream.h>
-#include <Encryption/WritableFile.h>
 
 #include <string>
