+++ dbms/src/Encryption/EncryptedWriteReadableFile.h
@@ -14,8 +14,8 @@
 
 #pragma once
 
+#include <BaseFile/WriteReadableFile.h>
 #include <Encryption/AESCTRCipherStream.h>
-#include <Encryption/WriteReadableFile.h>
 #include <common/types.h>
 
 #include <string>
