+++ dbms/src/Encryption/FileProvider.h
@@ -14,13 +14,13 @@
 
 #pragma once
 
+#include <BaseFile/RandomAccessFile.h>
+#include <BaseFile/WritableFile.h>
+#include <BaseFile/WriteReadableFile.h>
 #include <Core/Types.h>
 #include <Encryption/BlockAccessCipherStream.h>
 #include <Encryption/FileProvider_fwd.h>
 #include <Encryption/KeyManager.h>
-#include <Encryption/RandomAccessFile.h>
-#include <Encryption/WritableFile.h>
-#include <Encryption/WriteReadableFile.h>
 #include <Storages/Page/PageDefinesBase.h>
