+++ dbms/src/IO/ReadBufferFromRandomAccessFile.cpp
@@ -12,9 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+#include <BaseFile/RandomAccessFile.h>
 #include <Common/CurrentMetrics.h>
 #include <Common/ProfileEvents.h>
-#include <Encryption/RandomAccessFile.h>
 #include <IO/ReadBufferFromRandomAccessFile.h>
 
 namespace ProfileEvents
