+++ dbms/src/IO/WriteBufferFromWritableFile.cpp
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+#include <BaseFile/WritableFile.h>
 #include <Common/ProfileEvents.h>
-#include <Encryption/WritableFile.h>
 #include <IO/WriteBufferFromWritableFile.h>
