+++ dbms/src/Server/Server.cpp
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 #include <AggregateFunctions/registerAggregateFunctions.h>
+#include <BaseFile/RateLimiter.h>
 #include <Common/CPUAffinityManager.h>
 #include <Common/ComputeLabelHolder.h>
 #include <Common/Config/ConfigReloader.h>
@@ -40,7 +41,6 @@
 #include <Encryption/DataKeyManager.h>
 #include <Encryption/FileProvider.h>
 #include <Encryption/MockKeyManager.h>
-#include <Encryption/RateLimiter.h>
 #include <Flash/DiagnosticsService.h>
 #include <Flash/FlashService.h>
 #include <Flash/Mpp/GRPCCompletionQueuePool.h>
