+++ dbms/src/Server/StorageConfigParser.cpp
@@ -442,134 +442,6 @@ std::tuple<size_t, TiFlashStorageConfig> TiFlashStorageConfig::parseSettings(
     return std::make_tuple(global_capacity_quota, storage_config);
 }
 
-void StorageIORateLimitConfig::parse(const String & storage_io_rate_limit, const LoggerPtr & log)
-{
-    std::istringstream ss(storage_io_rate_limit);
-    cpptoml::parser p(ss);
-    auto config = p.parse();
-
-    readConfig(config, "max_bytes_per_sec", max_bytes_per_sec);
-    readConfig(config, "max_read_bytes_per_sec", max_read_bytes_per_sec);
-    readConfig(config, "max_write_bytes_per_sec", max_write_bytes_per_sec);
-    readConfig(config, "foreground_write_weight", fg_write_weight);
-    readConfig(config, "background_write_weight", bg_write_weight);
-    readConfig(config, "foreground_read_weight", fg_read_weight);
-    readConfig(config, "background_read_weight", bg_read_weight);
-    readConfig(config, "emergency_pct", emergency_pct);
-    readConfig(config, "high_pct", high_pct);
-    readConfig(config, "medium_pct", medium_pct);
-    readConfig(config, "tune_base", tune_base);
-    readConfig(config, "min_bytes_per_sec", min_bytes_per_sec);
-    readConfig(config, "auto_tune_sec", auto_tune_sec);
-
-    use_max_bytes_per_sec = (max_read_bytes_per_sec == 0 && max_write_bytes_per_sec == 0);
-
-    LOG_DEBUG(log, "storage.io_rate_limit {}", toString());
-}
-
-std::string StorageIORateLimitConfig::toString() const
-{
-    return fmt::format(
-        "max_bytes_per_sec {} max_read_bytes_per_sec {} max_write_bytes_per_sec {} use_max_bytes_per_sec {} "
-        "fg_write_weight {} bg_write_weight {} fg_read_weight {} bg_read_weight {} fg_write_max_bytes_per_sec {} "
-        "bg_write_max_bytes_per_sec {} fg_read_max_bytes_per_sec {} bg_read_max_bytes_per_sec {} emergency_pct {} "
-        "high_pct {} "
-        "medium_pct {} tune_base {} min_bytes_per_sec {} auto_tune_sec {}",
-        max_bytes_per_sec,
-        max_read_bytes_per_sec,
-        max_write_bytes_per_sec,
-        use_max_bytes_per_sec,
-        fg_write_weight,
-        bg_write_weight,
-        fg_read_weight,
-        bg_read_weight,
-        getFgWriteMaxBytesPerSec(),
-        getBgWriteMaxBytesPerSec(),
-        getFgReadMaxBytesPerSec(),
-        getBgReadMaxBytesPerSec(),
-        emergency_pct,
-        high_pct,
-        medium_pct,
-        tune_base,
-        min_bytes_per_sec,
-        auto_tune_sec);
-}
-
-UInt64 StorageIORateLimitConfig::readWeight() const
-{
-    return fg_read_weight + bg_read_weight;
-}
-
-UInt64 StorageIORateLimitConfig::writeWeight() const
-{
-    return fg_write_weight + bg_write_weight;
-}
-
-UInt64 StorageIORateLimitConfig::totalWeight() const
-{
-    return readWeight() + writeWeight();
-}
-
-UInt64 StorageIORateLimitConfig::getFgWriteMaxBytesPerSec() const
-{
-    if (writeWeight() <= 0 || totalWeight() <= 0)
-    {
-        return 0;
-    }
-    return use_max_bytes_per_sec ? static_cast<UInt64>(1.0 * max_bytes_per_sec / totalWeight() * fg_write_weight)
-                                 : static_cast<UInt64>(1.0 * max_write_bytes_per_sec / writeWeight() * fg_write_weight);
-}
-
-UInt64 StorageIORateLimitConfig::getBgWriteMaxBytesPerSec() const
-{
-    if (writeWeight() <= 0 || totalWeight() <= 0)
-    {
-        return 0;
-    }
-    return use_max_bytes_per_sec ? static_cast<UInt64>(1.0 * max_bytes_per_sec / totalWeight() * bg_write_weight)
-                                 : static_cast<UInt64>(1.0 * max_write_bytes_per_sec / writeWeight() * bg_write_weight);
-}
-
-UInt64 StorageIORateLimitConfig::getFgReadMaxBytesPerSec() const
-{
-    if (readWeight() <= 0 || totalWeight() <= 0)
-    {
-        return 0;
-    }
-    return use_max_bytes_per_sec ? static_cast<UInt64>(1.0 * max_bytes_per_sec / totalWeight() * fg_read_weight)
-                                 : static_cast<UInt64>(1.0 * max_read_bytes_per_sec / readWeight() * fg_read_weight);
-}
-
-UInt64 StorageIORateLimitConfig::getBgReadMaxBytesPerSec() const
-{
-    if (readWeight() <= 0 || totalWeight() <= 0)
-    {
-        return 0;
-    }
-    return use_max_bytes_per_sec ? static_cast<UInt64>(1.0 * max_bytes_per_sec / totalWeight() * bg_read_weight)
-                                 : static_cast<UInt64>(1.0 * max_read_bytes_per_sec / readWeight() * bg_read_weight);
-}
-
-UInt64 StorageIORateLimitConfig::getWriteMaxBytesPerSec() const
-{
-    return getBgWriteMaxBytesPerSec() + getFgWriteMaxBytesPerSec();
-}
-
-UInt64 StorageIORateLimitConfig::getReadMaxBytesPerSec() const
-{
-    return getBgReadMaxBytesPerSec() + getFgReadMaxBytesPerSec();
-}
-
-bool StorageIORateLimitConfig::operator==(const StorageIORateLimitConfig & config) const
-{
-    return config.max_bytes_per_sec == max_bytes_per_sec && config.max_read_bytes_per_sec == max_read_bytes_per_sec
-        && config.max_write_bytes_per_sec == max_write_bytes_per_sec && config.bg_write_weight == bg_write_weight
-        && config.fg_write_weight == fg_write_weight && config.bg_read_weight == bg_read_weight
-        && config.fg_read_weight == fg_read_weight && config.emergency_pct == emergency_pct
-        && config.high_pct == high_pct && config.medium_pct == medium_pct && config.tune_base == tune_base
-        && config.min_bytes_per_sec == min_bytes_per_sec && config.auto_tune_sec == auto_tune_sec;
-}
-
 void StorageS3Config::parse(const String & content)
 {
     std::istringstream ss(content);
