+++ dbms/src/Server/tests/gtest_storage_config.cpp
@@ -585,7 +585,7 @@ max_bytes_per_sec=1024000
 }
 CATCH
 
-TEST(StorageIORateLimitConfigTest, StorageIORateLimitConfig)
+TEST(IORateLimitConfigTest, IORateLimitConfig)
 try
 {
     Strings tests = {
@@ -637,7 +637,7 @@ background_read_weight=2
 
     auto log = Logger::get();
 
-    auto verify_default = [](const StorageIORateLimitConfig & io_config) {
+    auto verify_default = [](const IORateLimitConfig & io_config) {
         ASSERT_EQ(io_config.max_bytes_per_sec, 0);
         ASSERT_EQ(io_config.max_read_bytes_per_sec, 0);
         ASSERT_EQ(io_config.max_write_bytes_per_sec, 0);
@@ -655,7 +655,7 @@ background_read_weight=2
         ASSERT_EQ(io_config.getBgWriteMaxBytesPerSec(), 0);
     };
 
-    auto verify_case0 = [](const StorageIORateLimitConfig & io_config) {
+    auto verify_case0 = [](const IORateLimitConfig & io_config) {
         ASSERT_EQ(io_config.max_bytes_per_sec, 0);
         ASSERT_EQ(io_config.max_read_bytes_per_sec, 0);
         ASSERT_EQ(io_config.max_write_bytes_per_sec, 0);
@@ -673,7 +673,7 @@ background_read_weight=2
         ASSERT_EQ(io_config.getBgWriteMaxBytesPerSec(), 0);
     };
 
-    auto verify_case1 = [](const StorageIORateLimitConfig & io_config) {
+    auto verify_case1 = [](const IORateLimitConfig & io_config) {
         ASSERT_EQ(io_config.max_bytes_per_sec, 1024000);
         ASSERT_EQ(io_config.max_read_bytes_per_sec, 0);
         ASSERT_EQ(io_config.max_write_bytes_per_sec, 0);
@@ -691,7 +691,7 @@ background_read_weight=2
         ASSERT_EQ(io_config.getBgReadMaxBytesPerSec(), 102400 * 2);
     };
 
-    auto verify_case2 = [](const StorageIORateLimitConfig & io_config) {
+    auto verify_case2 = [](const IORateLimitConfig & io_config) {
         ASSERT_EQ(io_config.max_bytes_per_sec, 0);
         ASSERT_EQ(io_config.max_read_bytes_per_sec, 1024000);
         ASSERT_EQ(io_config.max_write_bytes_per_sec, 1024000);
@@ -709,7 +709,7 @@ background_read_weight=2
         ASSERT_EQ(io_config.getBgWriteMaxBytesPerSec(), 682666);
     };
 
-    auto verify_case3 = [](const StorageIORateLimitConfig & io_config) {
+    auto verify_case3 = [](const IORateLimitConfig & io_config) {
         ASSERT_EQ(io_config.max_bytes_per_sec, 1024000);
         ASSERT_EQ(io_config.max_read_bytes_per_sec, 1024000);
         ASSERT_EQ(io_config.max_write_bytes_per_sec, 1024000);
@@ -727,7 +727,7 @@ background_read_weight=2
         ASSERT_EQ(io_config.getBgWriteMaxBytesPerSec(), 102400 * 2);
     };
 
-    std::vector<std::function<void(const StorageIORateLimitConfig &)>> case_verifiers;
+    std::vector<std::function<void(const IORateLimitConfig &)>> case_verifiers;
     case_verifiers.push_back(verify_case0);
     case_verifiers.push_back(verify_case1);
     case_verifiers.push_back(verify_case2);
@@ -741,7 +741,7 @@ background_read_weight=2
         LOG_INFO(log, "parsing [index={}] [content={}]", i, test_case);
         ASSERT_TRUE(config->has("storage.io_rate_limit"));
 
-        StorageIORateLimitConfig io_config;
+        IORateLimitConfig io_config;
         verify_default(io_config);
         io_config.parse(config->getString("storage.io_rate_limit"), log);
         case_verifiers[i](io_config);
