+++ dbms/src/Storages/DeltaMerge/workload/Limiter.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include <Encryption/RateLimiter.h>
+#include <BaseFile/RateLimiter.h>
 #include <Storages/DeltaMerge/workload/Limiter.h>
 #include <Storages/DeltaMerge/workload/Options.h>
 #include <fmt/core.h>
