+++ dbms/src/Storages/Page/V3/tests/gtest_page_storage.cpp
@@ -12,12 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+#include <BaseFile/PosixRandomAccessFile.h>
+#include <BaseFile/RandomAccessFile.h>
+#include <BaseFile/RateLimiter.h>
 #include <Common/FailPoint.h>
 #include <Common/SyncPoint/Ctl.h>
 #include <Encryption/MockKeyManager.h>
-#include <Encryption/PosixRandomAccessFile.h>
-#include <Encryption/RandomAccessFile.h>
-#include <Encryption/RateLimiter.h>
 #include <Interpreters/Context.h>
 #include <Storages/Page/ConfigSettings.h>
 #include <Storages/Page/Page.h>
