+++ dbms/src/Storages/S3/MemoryRandomAccessFile.h
@@ -14,8 +14,8 @@
 
 #pragma once
 
+#include <BaseFile/RandomAccessFile.h>
 #include <Common/Exception.h>
-#include <Encryption/RandomAccessFile.h>
 #include <common/types.h>
 
 namespace DB
