+++ dbms/src/Storages/S3/S3RandomAccessFile.h
@@ -14,9 +14,9 @@
 
 #pragma once
 
+#include <BaseFile/RandomAccessFile.h>
 #include <Common/Exception.h>
 #include <Common/Logger.h>
-#include <Encryption/RandomAccessFile.h>
 #include <Storages/DeltaMerge/File/MergedFile.h>
 #include <Storages/DeltaMerge/ScanContext_fwd.h>
 #include <aws/s3/model/GetObjectResult.h>
