+++ dbms/src/Storages/S3/S3WritableFile.h
@@ -14,8 +14,8 @@
 
 #pragma once
 
+#include <BaseFile/WritableFile.h>
 #include <Common/Exception.h>
-#include <Encryption/WritableFile.h>
 #include <Storages/S3/S3Common.h>
 #include <common/types.h>
