+++ dbms/src/TestUtils/MockReadLimiter.h
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include <Encryption/RateLimiter.h>
+#include <BaseFile/RateLimiter.h>
 
 namespace DB
 {
